Implemented View Gathering.
You can now view the gatherings you have created and only that. Future work, need to create so that you can view the gatherings you are invited in and all gatherings both that you have created and the gatherings you are invited to.

Most was already ready but it did not work with skeleton. So i made kristians kode work with skeleton.